minor: updated datePickerWrapper
