+++ frontend/web/src/components/layout/createRent-DeviceEntry.component.ts
@@ -12,12 +12,12 @@ import {faCamera, faHelicopter, faLightbulb, faMicrophone, faXmark, faHashtag} f
 import AirDatepicker from 'air-datepicker';
 import {CreateRentComponent} from "./createRent.component"
 import {CreateRentDTO, RentTypeEnum} from "../../service/rent.service"
-import {Api, ccResponse, config, Regex} from "../../base"
+import {Api, ccResponse, config, DatePickerWrapper, Regex} from "../../base"
 import {AppState} from "../../AppState"
 import localeDe from "air-datepicker/locale/de"
 import {AutocompleteComponent, AutocompleteOption} from "../basic/autocomplete.component"
 import DeviceTypeService, {DeviceType, DeviceTypeSource, DeviceTypeVariantEnum} from "../../service/deviceType.service"
-import Util, {AnimationHelper, DatePickerWrapper, Logger} from "../../util"
+import Util, {AnimationHelper, Logger} from "../../util"
 import DeviceService, {Device, DeviceDTO} from "../../service/device.service"
 
 export interface CreateRentDeviceEntryData {
