+++ frontend/web/styles/components/layout/rentListEntry.styles.scss
@@ -47,7 +47,7 @@ input, cc-autocomplete{
 
 //TODO format all dates with more letter spacing as they are in figma
 
-div{//TODO cleanup this selector
+.area{//TODO cleanup this selector
   display: flex;
   align-items: center;
   gap: 0.1rem;
